29-gemini3/pyproject.toml
[project]
name = "gemini3-demo"
version = "0.1.0"
description = "Gemini 3 demo script"
readme = "GEMINI.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
]
